requirements.txt
streamlit
opencv-python
numpy
pandas
seaborn
pytesseract
albumentations
easyocr
scikit-image
pillow
timm
PaddleOCR
EasyOCR 